service에서 게시글 수정·삭제 권한 체크 로직 구현

+++ src/main/java/com/lsj/demo/controller/UserArticleController.java
@@ -60,12 +60,19 @@ public class UserArticleController {
 		if (article == null) {
 			return ResultData.from("F-1", Ut.f("%d번 글이 존재하지 않습니다.", id));
 		}
+		
+		ResultData loginedMemberCanModifyRd = articleService.loginedMemberCanModify(loginedMemberId, article);
+
+		if (loginedMemberCanModifyRd.isFail()) {
+			return ResultData.from(loginedMemberCanModifyRd.getResultCode(), loginedMemberCanModifyRd.getMsg());
+		}
 
 		articleService.modifyArticle(id, title, body);
 		article = articleService.getArticleById(id);
 
-		return ResultData.from("S-1", Ut.f("%d번 게시글이 수정되었습니다.", id), article);
+		return ResultData.from(loginedMemberCanModifyRd.getResultCode(), loginedMemberCanModifyRd.getMsg(), article);
 	}
+
 
 	@RequestMapping("/usr/article/doDelete")
 	@ResponseBody
@@ -87,6 +94,10 @@ public class UserArticleController {
 		
 		if (article == null) {
 			return ResultData.from("F-1", Ut.f("%d번 글이 존재하지 않습니다.", id));
+		}
+		
+		if (article.getMemberId() != loginedMemberId) {
+			return ResultData.from("F-A2", "해당 게시글에 대한 권한이 존재하지 않습니다.");
 		}
 
 		articleService.deleteArticle(id);
